02_001: change the file structure of the classes in Model
