+++ wetherForecastApp/Models/WeatherParameters.cs
