+++ wetherForecastApp/Properties/AssemblyInfo.cs
