+++ wetherForecastApp/Scripts/jquery.validate.unobtrusive.js
